feat: let users use the console without an account
Importing a space from someone else is a totally valid way to use the service, which means we can let users in to kick the tires before they have an account.

To enable this I moved the gates around, tweaked styling, and handled a couple UX cases where users might have seen empty space lists and such.

I think this is way nicer, but it could probably use a little of that @olizilla design sense.