fix: delegate space recovery after creation (#58)
We were not doing this, and as a result any spaces created in the web
console recently will not be recoverable.

I'd like to get this bugfix into production ASAP but have left TODOs for
follow-on work that should be completed as soon as possible.

We should also give users the ability to see if their space is
recoverable from the service. I'll create issues for all of this.