wip: tell users when they have unrecoverable spaces
Use a heuristic to look at available claims and determine whether a space is recoverable. Currently considers a space recoverable if any abilities are delegated and should probably validate a minimum set.

Unfortunately we don't have any way for users to recover an "unrecoverable" space at the moment, since we still discard Space keys after creation - I'm not convinced this feature is particularly helpful without some way to fix unrecoverable spaces, but it is probably better than not having any way to determine that at all!