feat: directory uploads
Use the new w3ui Uploader's directory upload support to allow users to
upload directories. Also allow them to wrap a single file in a directory.

I've moved this to a radio group because we are going to add the ability
to upload a CAR file, which will be a third option and will use a
different upload function.