feat: show a preview of the url in the cid page.
Very cheap preview of the gateway view in the cid page via an iframe.

Disable interaction on the iframe with the `inert` attributes. No clicking through, that way lies madness.

We probably don't want to do this, but it does show how useful a visual preview of the contnet would be.

License: MIT
Signed-off-by: Oli Evans <oli@protocol.ai>