fix: tweak plan page styles
there was an edge case right at one of the most common sizes, so:

a) use `nbsp`s to make sure the words in the usage bar never go to two levels
b) fine tune the `lg` breakpoint on the plan page

I'm using a custom width `w-[21rem]` that is still in-line with the design system but imho should not be used more widely so I've decided _not_ to add it to the tailwind config. If we end up using this in more places it should be added there later.