fix: Revert "fix: Revert "feat: Use Blob protocol"" (😩) (#136)
Reverts storacha/console#134 (which reverted
https://github.com/storacha/console/pull/132)

Last night, this was still failing locally. This morning it's working
again. I've done nothing but investigate.

I now think the Blob protocol was a red herring. I think something else
caused us to send a receipt that the client chokes on, and that that'll
happen regardless of this change. If it comes back, we should
investigate further, but I don't think this is what did it.