fix: Use Ucanto 10 across the board (#140)
[Ucanto 10 added support for embedded
effects.](https://github.com/storacha/ucanto/pull/347) That's why
uploads have been failing: we're still on 9 in some places.

The whole way we load things here needs an overhaul: we have multiple
instances of some packages where it only makes sense to have one, and to
piece things together as peer dependencies. But that's a huge thing to
solve, so for now, brute force it by just making sure everything is up
to date.