fix: I did React wrong
I was thinking the state would be shared between invocations of the hook and that's just not the case! I'd need to introduce a context to get it to work the way I expected. DOH.

move to an argument passing style because that's easier than introducing a context, for now